Refactor: FlipCard 컴포넌트 Back, Face 분리 [#15]
